fix: new boilerplate tests
